Added OpenNearbyDoors action, added NavMeshObstacles for intoxication traps, got all Labyrinth chambers working
