+++ Client/BehaviorExtensions/CustomLayerDelayedUpdate.cs
@@ -24,7 +24,8 @@ namespace QuestingBots.BehaviorExtensions
         Sleep,
         ToggleSwitch,
         UnlockDoor,
-        CloseNearbyDoors
+        CloseNearbyDoors,
+        OpenNearbyDoors
     }
 
     internal abstract class CustomLayerDelayedUpdate : CustomLayer
@@ -56,8 +57,6 @@ namespace QuestingBots.BehaviorExtensions
 
         public override Action GetNextAction()
         {
-            //Singleton<LoggingUtil>.Instance.LogInfo(BotOwner.GetText() + " is swtiching from " + previousAction.ToString() + " to " + nextAction.ToString());
-
             previousAction = nextAction;
 
             switch (nextAction)
@@ -75,6 +74,7 @@ namespace QuestingBots.BehaviorExtensions
                 case BotActionType.ToggleSwitch: return new Action(typeof(BotLogic.Objective.ToggleSwitchAction), actionReason);
                 case BotActionType.UnlockDoor: return new Action(typeof(BotLogic.Objective.UnlockDoorAction), actionReason);
                 case BotActionType.CloseNearbyDoors: return new Action(typeof(BotLogic.Objective.CloseNearbyDoorsAction), actionReason);
+                case BotActionType.OpenNearbyDoors: return new Action(typeof(BotLogic.Objective.OpenNearbyDoorsAction), actionReason);
             }
 
             throw new InvalidOperationException("Invalid action selected for layer");
